requirements.txt
Django==4.2.4
djangorestframework==3.14.0
psycopg2>=2.8.6,<2.9